Create README.md
"This project is an Asp.Net Core 5.0 application. It is developed within the scope of Asp.Net Core 5.0. The project focuses on building a Dynamic MyCV Blog. It is implemented using C# programming language, Sql Server 2019, Entity Framework Core, and follows the N-Tier Architecture. Code-First approach is utilized, and for frontend design, HTML, CSS, and Bootstrap are used.

The sections for retrieving data from the web through APIs have been completed. As for the messaging functionalities and Ajax, they will be added gradually by me.

Thank you!"